new static methods to instantiate SMPTE objects
